requirements.txt
# Skippy SAP EWM Assistant - Production Requirements
# Dependencies for skippy_coach_production.py, config.py, and build_index.py
# Version: 1.0.0

# Core Framework (required for streamlit app)
streamlit>=1.28.0

# LangChain Core Components (used in production.py and build_index.py)
langchain-openai>=0.0.2
langchain-chroma>=0.1.2
langchain-core>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.38

# ChromaDB Vector Database
chromadb>=0.4.15

# Configuration Management (used in config.py)
pydantic>=2.0.0
python-dotenv>=1.0.0

# PDF Processing and Text Splitting (used in build_index.py)
PyMuPDF>=1.23.0
tiktoken>=0.5.1

# Fallback embedding support (used in build_index.py)
sentence-transformers>=2.2.2
